pyproject.toml
[tool.poetry]
name = "third-pro"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
pymysql = "^1.1.0"
cryptography = "^42.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
